requirements.txt
# ── Web Framework ─────────────────────────────────────────────
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9

# ── Telegram Bot ──────────────────────────────────────────────
python-telegram-bot==21.3

# ── AI / LLM (Free Groq API) ──────────────────────────────────
groq==0.9.0

# ── Local Embeddings & Vector Search (100% free, offline) ─────
sentence-transformers==3.0.1
chromadb==0.5.3

# ── Database ──────────────────────────────────────────────────
sqlalchemy==2.0.31
aiosqlite==0.20.0

# ── HTTP Client ───────────────────────────────────────────────
httpx==0.27.0

# ── Environment & Config ──────────────────────────────────────
python-dotenv==1.0.1

# ── Data Validation ───────────────────────────────────────────
pydantic==2.8.2
pydantic-settings==2.3.4

# ── Utilities ─────────────────────────────────────────────────
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

gunicorn==21.2.0